Added Date_Helper class and copied its single function, datetime_from_input, to validation.php as well because it's also used for validation.
The datetime_from_input function is extremely useful when you have to support a bunch of datetime formats. PHP does not make datetime validation easy, so this is far from perfect, but it should get you most of the way to a proper solution.

I still have more coming later today.